Really stupid idea for a bugfix:
the huddle_estimators in the ram branch and main are identical. But the one in main has a more recent "change". This may be the reason why I get a "Please commit your changes or stash them before you switch branches" error for this specific file when trying to pip install from the branch which then prevents the installation, even though there literally is no changes that even could be committed and the folder I install into was absolutely empty before.

So I first break the huddle estimator, then restore it, to have the change on the branch be more recent than the one on main.